Moved recursive hasher CLI functionality into helper
