+++ examples/list_file_entries.py
@@ -65,12 +65,12 @@ class FileEntryLister(object):
       magnitude_1024 += 1
 
     size_string_1000 = None
-    if 0 <= magnitude_1000 <= 7:
+    if 0 < magnitude_1000 <= 7:
       size_string_1000 = '{0:.1f}{1:s}'.format(
           size_1000, self._UNITS_1000[magnitude_1000])
 
     size_string_1024 = None
-    if 0 <= magnitude_1024 <= 7:
+    if 0 < magnitude_1024 <= 7:
       size_string_1024 = '{0:.1f}{1:s}'.format(
           size_1024, self._UNITS_1024[magnitude_1024])
